% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodev.R
\docType{data}
\name{biodev}
\alias{biodev}
\title{Base biodev class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
biodev
}
\description{
Initializes a common information required for all docker-based biodev tools.
}
\keyword{datasets}
