% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodevtools}
\alias{biodevtools}
\alias{biodevtools-package}
\title{Support classes to use bioinformatic tools in R using Docker}
\description{
Support classes to use bioinformatic tools in R using Docker
}
